Derive Macro validation_macros::ValidatingType

#[derive(ValidatingType)]
{
    // Attributes available to this derive:
    #[field_type]
    #[validate]
}

Scaffolds validation and provides a validate function on the unvalidated type by implementing validation::ValidatingType

Example

// Given you have this struct: struct Message { cpu_percentage: i32 }

// Create this struct (must be named the same as the unvalidated type with the prefix ‘Validated’): #[derive(validation_macros::ValidatingType)] struct ValidatedMessage { #[field_type(i32)] cpu_percentage: u8 }

The macro will then generate a trait MessageTypeValidator and an empty struct MessageValidator. You must impl MessageTypeValidator for MessageValidator.

Decorate fields with the field_type attribute, when the unvalidated type differs from the validated type. See example above.

Optionally, decorate fields with the validate attribute to generate default implementations:

#[validate] will call ValidatedFieldType::validate with the input automatically wrapped in Some

#[validate(opt)] will call ValidatedFieldType::validate_optional

#[validate(none)] will pass through the input without performing any validation (input and output type must be the same)